+++ packages/jaspr_bootstrap/example/lib/pages/carousel_page.dart
@@ -1,5 +1,6 @@
 import 'package:jaspr/jaspr.dart';
-import 'package:jaspr_bootstrap/jaspr_bootstrap.dart';
+import 'package:jaspr_ui/bootstrap.dart';
+import 'package:jaspr_ui/core.dart';
 
 const image1 =
     'https://media.istockphoto.com/photos/abstract-wavy-object-picture-id1198271727?b=1&k=20&m=1198271727&s=170667a&w=0&h=b626WM5c-lq9g_yGyD0vgufb4LQRX9UgYNWPaNUVses=';
