+++ packages/jaspr_bootstrap/example/lib/pages/first_page.dart
@@ -1,5 +1,6 @@
 import 'package:jaspr/jaspr.dart';
-import 'package:jaspr_bootstrap/jaspr_bootstrap.dart';
+import 'package:jaspr_ui/bootstrap.dart';
+import 'package:jaspr_ui/core.dart';
 
 class App extends StatelessComponent {
   @override
