--- packages/jaspr_bootstrap/lib/jaspr_bootstrap.dart
@@ -1,19 +1,0 @@
-library jaspr_bootstrap;
-
-export 'src/core/page_style.dart';
-export 'src/components/carousel.dart';
-export 'src/layout/center.dart';
-export 'src/layout/container.dart';
-export 'src/layout/column.dart';
-export 'src/layout/row.dart';
-export 'src/layout/size.dart';
-export 'src/elements/button.dart';
-export 'src/elements/header.dart';
-export 'src/elements/horizontal_line.dart';
-export 'src/elements/image.dart';
-export 'src/elements/paragraph.dart';
-export 'src/edge_insets.dart';
-export 'src/enums.dart';
-export 'src/border.dart';
-
-
