+++ packages/jaspr_bootstrap/lib/src/bootstrap/layout/column.dart
@@ -1,44 +1,48 @@
 import 'package:jaspr/jaspr.dart';
-import 'package:jaspr_bootstrap/src/border.dart';
-import 'package:jaspr_bootstrap/src/edge_insets.dart';
-import 'package:jaspr_bootstrap/src/enums.dart';
+import 'package:jaspr_ui/bootstrap.dart';
 
-class Container extends StatelessComponent {
-  final Breakpoint breakpoint;
+class Column extends StatelessComponent {
   final BackgroundColor? backgroundColor;
   final TextColor? textColor;
   final EdgeInsets? padding;
   final EdgeInsets? margin;
   final Border? border;
-  final Map<String, String>? styles;
 
   final Component? _child;
   final List<Component>? _children;
+  final Flex? _flex;
+  final List<Flex>? _flexibility;
 
-  Container({
+  const Column({
     Key? key,
     Component? child,
     List<Component>? children,
-    this.breakpoint = Breakpoint.extraSmall,
+    Flex? flex,
+    List<Flex>? flexibility,
     this.backgroundColor,
     this.textColor,
     this.padding,
     this.margin,
     this.border,
-    this.styles,
   })  : _child = child,
         _children = children,
+        _flex = flex,
+        _flexibility = flexibility,
         super(key: key);
 
   List<Component> get children => [if (_child != null) _child!, ..._children ?? []];
+
+  List<Flex> get flexibility {
+    final List<Flex> result = [if (_flex != null) _flex!, ..._flexibility ?? []];
+    return result.isEmpty ? [Flex()] : result;
+  }
 
   @override
   Iterable<Component> build(BuildContext context) sync* {
     yield DomComponent(
       tag: 'div',
-      styles: styles,
       classes: [
-        'container${breakpoint.value}',
+        ...flexibility.map((e) => e.getClass('col')),
         if (backgroundColor != null) backgroundColor!.value,
         if (textColor != null) textColor!.value,
         if (padding != null) ...padding!.getClasses('p'),
@@ -48,4 +52,16 @@ class Container extends StatelessComponent {
       children: children,
     );
   }
+}
+
+class Flex {
+  final Breakpoint breakpoint;
+  final ColumnSpace space;
+
+  const Flex({
+    this.breakpoint = Breakpoint.extraSmall,
+    this.space = ColumnSpace.auto,
+  });
+
+  String getClass(String type) => space == ColumnSpace.auto ? type : '$type${breakpoint.value}-${space.value}';
 }
