+++ packages/jaspr_bootstrap/lib/src/bootstrap/layout/container.dart
@@ -1,27 +1,29 @@
 import 'package:jaspr/jaspr.dart';
-import 'package:jaspr_bootstrap/src/border.dart';
-import 'package:jaspr_bootstrap/src/edge_insets.dart';
-import 'package:jaspr_bootstrap/src/enums.dart';
+import 'package:jaspr_ui/bootstrap.dart';
 
-class Row extends StatelessComponent {
+class Container extends StatelessComponent {
+  final Breakpoint breakpoint;
   final BackgroundColor? backgroundColor;
   final TextColor? textColor;
   final EdgeInsets? padding;
   final EdgeInsets? margin;
   final Border? border;
+  final Map<String, String>? styles;
 
   final Component? _child;
   final List<Component>? _children;
 
-  Row({
+  Container({
     Key? key,
     Component? child,
     List<Component>? children,
+    this.breakpoint = Breakpoint.extraSmall,
     this.backgroundColor,
     this.textColor,
     this.padding,
     this.margin,
     this.border,
+    this.styles,
   })  : _child = child,
         _children = children,
         super(key: key);
@@ -32,8 +34,9 @@ class Row extends StatelessComponent {
   Iterable<Component> build(BuildContext context) sync* {
     yield DomComponent(
       tag: 'div',
+      styles: styles,
       classes: [
-        'row',
+        'container${breakpoint.value}',
         if (backgroundColor != null) backgroundColor!.value,
         if (textColor != null) textColor!.value,
         if (padding != null) ...padding!.getClasses('p'),
