+++ packages/jaspr_bootstrap/pubspec.yaml
@@ -1,4 +1,4 @@
-name: jaspr_bootstrap
+name: jaspr_ui
 description: Bootstrap components for jaspr.
 version: 0.1.0
